mic. conf. moved to its own type file
